+++ CHANGES.md
@@ -54,6 +54,7 @@
     claimed ownership of the file extension, e.g. VS2017 handling .dds files.
     The tab is now registered to handle all files, though it will only show
     up if the file is actually supported by rom-properties.
+  * DreamcastSave: Trim spaces from the end of text fields.
 
 * Other changes:
   * libromdata/ has been reorganized to use subdirectories for type of system.